+++ app/src/main/java/com/hongmyeoun/goldcalc/view/home/topbar/ProgressBar.kt
@@ -20,10 +20,10 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.hongmyeoun.goldcalc.model.common.toPercentage
 import com.hongmyeoun.goldcalc.ui.theme.CharacterEmblemBG
-import com.hongmyeoun.goldcalc.viewModel.main.CharacterListVM
+import com.hongmyeoun.goldcalc.viewModel.home.HomeVM
 
 @Composable
-fun ProgressBar(viewModel: CharacterListVM) {
+fun ProgressBar(viewModel: HomeVM) {
     val animatedProgress = animateFloatAsState(
         targetValue = viewModel.progressPercentage,
         animationSpec = ProgressIndicatorDefaults.ProgressAnimationSpec,
