+++ app/src/main/java/com/hongmyeoun/goldcalc/view/homework/HomeworkView.kt
@@ -22,11 +22,11 @@ import com.hongmyeoun.goldcalc.ui.theme.ImageBG
 import com.hongmyeoun.goldcalc.view.common.LoadingScreen
 import com.hongmyeoun.goldcalc.view.homework.bottomBar.HomeworkBottomBar
 import com.hongmyeoun.goldcalc.view.homework.content.HomeworkContent
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.AbyssDungeonVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.CommandBossVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.EpicRaidVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.HomeworkVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.KazerothRaidVM
+import com.hongmyeoun.goldcalc.viewModel.homework.AbyssDungeonVM
+import com.hongmyeoun.goldcalc.viewModel.homework.CommandBossVM
+import com.hongmyeoun.goldcalc.viewModel.homework.EpicRaidVM
+import com.hongmyeoun.goldcalc.viewModel.homework.HomeworkVM
+import com.hongmyeoun.goldcalc.viewModel.homework.KazerothRaidVM
 import kotlinx.coroutines.delay
 
 @Composable
