+++ app/src/main/java/com/hongmyeoun/goldcalc/view/homework/content/checkLists/CheckLists.kt
@@ -11,11 +11,11 @@ import com.hongmyeoun.goldcalc.view.homework.content.checkLists.raidCard.raidCon
 import com.hongmyeoun.goldcalc.view.homework.content.checkLists.raidCard.raidContents.Command
 import com.hongmyeoun.goldcalc.view.homework.content.checkLists.raidCard.raidContents.Epic
 import com.hongmyeoun.goldcalc.view.homework.content.checkLists.raidCard.raidContents.Kazeroth
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.AbyssDungeonVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.CommandBossVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.EpicRaidVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.HomeworkVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.KazerothRaidVM
+import com.hongmyeoun.goldcalc.viewModel.homework.AbyssDungeonVM
+import com.hongmyeoun.goldcalc.viewModel.homework.CommandBossVM
+import com.hongmyeoun.goldcalc.viewModel.homework.EpicRaidVM
+import com.hongmyeoun.goldcalc.viewModel.homework.HomeworkVM
+import com.hongmyeoun.goldcalc.viewModel.homework.KazerothRaidVM
 
 @Composable
 fun CheckLists(
